MatchMe/requirements.txt
Appium-Python-Client==0.26
selenium==3.0.2
xmlrunner==1.7.7
altunityrunner==0.2
pytest
pytest-allure-adaptor==1.7.10
